pyproject.toml
[project]
name = "icd-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "blinker==1.9.0",
    "cachetools==6.1.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "distro==1.9.0",
    "flask==3.1.1",
    "flask-cors==6.0.1",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.4.1",
    "markupsafe==3.0.2",
    "narwhals==2.1.2",
    "numpy==2.3.2",
    "openai==1.100.1",
    "packaging==25.0",
    "pandas==2.3.1",
    "pillow==11.3.0",
    "protobuf==6.32.0",
    "pyarrow==21.0.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydeck==0.9.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "referencing==0.36.2",
    "requests==2.32.5",
    "rpds-py==0.27.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "streamlit==1.48.1",
    "tenacity==9.1.2",
    "toml==0.10.2",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "watchdog==6.0.0",
    "werkzeug==3.1.3",
]
